requirements.txt
amqp==5.3.1
ansible==12.1.0
ansible-core==2.19.3
asgiref==3.9.2
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.2
Django==5.2.6
django-ratelimit==4.1.0
idna==3.10
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.3
mod_wsgi==5.0.2
mysqlclient==2.2.7
packaging==25.0
prompt_toolkit==3.0.52
pycparser==2.23
pyspnego==0.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyvmomi==9.0.0.0
pywinrm==0.5.0
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
requests-credssp==2.0.0
requests_ntlm==1.3.0
resolvelib==1.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
xmltodict==1.0.2
